Resultページにクエリを読み込む処理を追加
result-waitingページから送られた結果の情報が入ったクエリパラメータを読み取るようにしました。
結果の情報を表示する処理はまだ書いていません。
